requirements.txt
streamlit
qrcode[pil]
python-dotenv
streamlit-cookies-manager
